requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fonttools==4.58.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
narwhals==1.40.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
plotly==6.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyforest==1.1.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9.1
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
xlrd==2.0.1
